shared library

+++ vars/App.groovy
@@ -1,25 +1,34 @@
-def call(String tfaction) {
+def call(String action) {
     //define a String variable, s must be capital (S) in String.  
     pipeline {
         agent {
             stages {
                 stage ('Terraform apply') {
                     steps {
+                        if (action == 'apply'){
                         sh "terraform apply --auto-approve"
+                    }    
+                    }
                     }
                 stage ('Terraform init') {
                     steps {
+                        if (action == 'init'){
                         sh "Terraform init"
+                    }
                     }
                     }
                 stage ('Terraform destroy') {
                     steps {
+                    if (action == 'destroy'){
                         sh "terraform destroy --auto-approve"
+                    }
                     }
                     }
                 stage ('Terraform plan') {
                     steps {
+                        if (tfaction == 'plan') {
                         sh "terraform plan"
+                    }
                     }
                     }
                 }
